+++ tests/unit/models/ComponentDefinition.spec.js
@@ -11,8 +11,6 @@ describe('Test class: ComponentDefinition', () => {
       expect(componentDefinition.parentTypes).toEqual([]);
       expect(componentDefinition.definedAttributes).toEqual([]);
       expect(componentDefinition.isContainer).toBeFalsy();
-      expect(componentDefinition.displayable).toBeTruthy();
-      expect(componentDefinition.required).toBeFalsy();
     });
 
     it('Check passing undefined variables to constructor', () => {
@@ -24,8 +22,6 @@ describe('Test class: ComponentDefinition', () => {
       expect(componentDefinition.parentTypes).toEqual([]);
       expect(componentDefinition.definedAttributes).toEqual([]);
       expect(componentDefinition.isContainer).toBeFalsy();
-      expect(componentDefinition.displayable).toBeTruthy();
-      expect(componentDefinition.required).toBeFalsy();
     });
 
     it('Check passing variable to constructor', () => {
@@ -36,8 +32,6 @@ describe('Test class: ComponentDefinition', () => {
         parentTypes: ['type'],
         definedAttributes: ['attribute'],
         isContainer: true,
-        displayable: false,
-        required: true,
       });
 
       expect(componentDefinition.type).toEqual('type');
@@ -48,8 +42,6 @@ describe('Test class: ComponentDefinition', () => {
       expect(componentDefinition.definedAttributes.length).toEqual(1);
       expect(componentDefinition.definedAttributes[0]).toEqual('attribute');
       expect(componentDefinition.isContainer).toBeTruthy();
-      expect(componentDefinition.displayable).toBeFalsy();
-      expect(componentDefinition.required).toBeTruthy();
     });
   });
 });